requirements.txt
aniso8601==7.0.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
cfgv==3.4.0
channels==4.3.1
channels_redis==4.2.1
colorama==0.4.6
constantly==23.10.4
coverage==7.10.6
cryptography==46.0.1
daphne==2.5.0
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.6
django-environ==0.12.0
django-filter==25.1
django-graphql-jwt==0.4.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==23.0.0
hyperlink==21.0.0
identify==2.6.14
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
msgpack==1.1.1
nodeenv==1.9.1
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
promise==2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
Rx==1.6.3
service-identity==24.2.0
setuptools==80.9.0
singledispatch==4.1.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.34.0
whitenoise==6.11.0
zope.interface==8.0
